+++ TDEngine2/include/math/TVector3.h
@@ -233,4 +233,6 @@ namespace TDEngine2
 
 
 	template <> struct GetTypeId<TVector3> { TDE2_API TDE2_STATIC_CONSTEXPR TypeId mValue = TDE2_TYPE_ID(TVector3); };
+
+	template <> TDE2_API E_RESULT_CODE Serialize<TVector3>(class IArchiveWriter* pWriter, TVector3&& value);
 }