+++ TDEngine2/include/math/TVector4.h
@@ -193,4 +193,14 @@ namespace TDEngine2
 	*/
 
 	TDE2_API TVector4 Max(const TVector4& lvec4, const TVector4& rvec4);
+
+
+	/*!
+		\brief TVector4's Serialization/Deserialization helpers
+	*/
+
+	TDE2_API TResult<TVector4> LoadVector4(IArchiveReader* pReader);
+	TDE2_API E_RESULT_CODE SaveVector4(IArchiveWriter* pWriter, const TVector4& object);
+
+	template <> TDE2_API E_RESULT_CODE Serialize<TVector4>(class IArchiveWriter* pWriter, TVector4&& value);
 }