+++ TDEngine2/plugins/D3D11GraphicsContext/source/CD3D11ShaderCompiler.cpp
@@ -162,6 +162,7 @@ namespace TDEngine2
 						}
 
 						variableDesc.mSize = std::max<USIZE>(variableDesc.mSize, static_cast<USIZE>(shaderVariableInfoDesc.Size));
+						variableDesc.mOffset = shaderVariableInfoDesc.StartOffset;
 					}
 				}
 			}