+++ TDEngine2/source/math/TVector3.cpp
@@ -241,4 +241,10 @@ namespace TDEngine2
 
 		return RC_OK;
 	}
+
+
+	template <> TDE2_API E_RESULT_CODE Serialize<TVector3>(IArchiveWriter* pWriter, TVector3&& value)
+	{
+		return pWriter->SetUInt32("type_id", static_cast<U32>(GetTypeId<TVector3>::mValue)) | SaveVector3(pWriter, value);
+	}
 }